requirements.txt
torch==1.13.1
torchvision==0.14.1
streamlit==1.19.0
pillow==9.4.0
numpy==1.24.2
scikit-learn==1.2.1
matplotlib==3.7.1
seaborn==0.12.2
